testtask/requirements.txt
django
djangorestframework
uvicorn
whitenoise
pytest
pytest-django
pydantic
email-validator
pyTelegramBotAPI